refactored interface
